==> 08_figures.Rmd <==
---
title: "08_figures"
author: "Flora Hann"
date: "2024-11-07"
output: html_document
---

# setup and loading packages
```{r}
options(scipen = 999, digits = 15)

suppressPackageStartupMessages(library(rstatix))
suppressPackageStartupMessages(library(janitor))
suppressPackageStartupMessages(library(plotrix))
suppressPackageStartupMessages(library(showtext))
suppressPackageStartupMessages(library(extrafont))
suppressPackageStartupMessages(library(tidyverse))

loadfonts(device = "all")
font_add(family = "Arial", regular = "fonts/Arial.ttf")
showtext_auto()
```

# importing data
```{r}
# ASRT
asrt_merged_rt <- read_csv("processed_data/asrt_merged_rt.csv", show_col_types = F) %>% mutate(aq_group = factor(aq_group, levels = c(1, 2), labels = c("Lower AQ", "Higher AQ")))
asrt_merged_acc <- read_csv("processed_data/asrt_merged_acc.csv", show_col_types = F) %>% mutate(aq_group = factor(aq_group, levels = c(1, 2), labels = c("Lower AQ", "Higher AQ")))
rt_acc_binwise <- read_csv("processed_data/rt_acc_binwise.csv", show_col_types = F) %>% mutate(aq_group = factor(aq_group, levels = c(1, 2), labels = c("Lower AQ", "Higher AQ")))
```

# Figure 1: AQ distribution
```{r}
bar_chart_aq_distribution <- asrt_merged_rt %>% 
  select(id, aq_sum) %>% 
  distinct() %>% 
  count(aq_sum) %>% 
  ggplot(aes(x = aq_sum, y = n, fill = aq_sum)) +
  geom_col() +
  scale_fill_gradient(low = "#F3C22D", high = "#3C96DF") +
  labs(x = "AQ total score", y = "Number of participants") +
  theme_classic(base_size = 18) +
  theme(
    legend.title = element_blank()
  )

bar_chart_aq_distribution
```

# Figure 3: Statistical learning as a function of AQ - reaction time
```{r}
# preprocessing
asrt_merged_rt_learning <- asrt_merged_rt %>% 
  select(id, aq_sum, aq_group, bin, ls_binwise_rt) %>% 
  filter(bin == 1 | bin == 2 | bin == 3) %>%
  distinct()
```

## Panel A: line plots with lower and higher AQ groups
```{r}
line_plot_rt_learning <- asrt_merged_rt_learning %>%
  group_by(bin, aq_group) %>% 
  summarise(mean_ls_rt = mean(ls_binwise_rt, na.rm = T), se_rt = std.error(ls_binwise_rt, na.rm = T)) %>% 
  ggplot(aes(bin, mean_ls_rt, color = aq_group)) +
  geom_point() +
  geom_line(linewidth = 1.5) +
  geom_errorbar(aes(ymin = mean_ls_rt - se_rt, ymax = mean_ls_rt + se_rt), width = 0.1, linewidth = 1) +
  theme_classic(base_size = 18) +
  scale_color_manual(values = c("#F3C22D", "#3C96DF")) +
  theme(
    legend.title = element_blank()
  ) +
  labs(y = "RT learning score (ms)", x = "Bin") +
  scale_x_continuous(breaks = seq(1, 3, 1))

line_plot_rt_learning
```

## Panel B: binwise correlations between AQ and learning scores
```{r}
scatter_plot_rt_learning <- asrt_merged_rt_learning %>% 
  ggplot(aes(aq_sum, ls_binwise_rt, color = aq_sum)) +
  geom_jitter() +
  geom_smooth(method = lm, color = "black", lwd = 0.6, alpha = 1) +
  facet_wrap(~ bin) +
  scale_color_gradient(low = "#F3C22D", high = "#3C96DF") +
  theme_classic(base_size = 18) +
  scale_y_continuous(limits = c(-60, 60)) +
  scale_x_continuous(limits = c(NA, 40)) +
  labs(y = "RT learning score (ms)", x = "AQ score", color = "AQ score")

scatter_plot_rt_learning
```

# Figure 4: Statistical learning as a function of AQ - accuracy
```{r}
# preprocessing
asrt_merged_acc_learning <- asrt_merged_acc %>% 
  select(id, aq_sum, aq_group, bin, ls_binwise_acc) %>% 
  filter(bin == 1 | bin == 2 | bin == 3) %>%
  distinct()
```

## Panel A: line plots with lower and higher AQ groups
```{r}
line_plot_acc_learning <- asrt_merged_acc_learning %>%
  group_by(bin, aq_group) %>% 
  summarise(mean_ls_acc = mean(ls_binwise_acc, na.rm = T), se_acc = std.error(ls_binwise_acc, na.rm = T)) %>% 
  ggplot(aes(bin, 100 * mean_ls_acc, color = aq_group)) +
  geom_point() +
  geom_line(linewidth = 1.5) +
  geom_errorbar(aes(ymin = 100 * mean_ls_acc - 100 * se_acc, ymax = 100 * mean_ls_acc + 100 * se_acc), width = 0.1, linewidth = 1) +
  theme_classic(base_size = 18) +
  scale_color_manual(values = c("#F3C22D", "#3C96DF")) +
  theme(
    legend.title = element_blank()
  ) +
  labs(y = "ACC learning score (%)", x = "Bin") +
  scale_x_continuous(breaks = seq(1, 3, 1))

line_plot_acc_learning
```

## Panel B: binwise correlations between AQ and learning scores
```{r}
scatter_plot_acc_learning <- asrt_merged_acc_learning %>% 
  ggplot(aes(aq_sum, 100 * ls_binwise_acc, color = aq_sum)) +
  geom_jitter() +
  geom_smooth(method = lm, color = "black", lwd = 0.6, alpha = 1) +
  facet_wrap(~ bin) +
  scale_color_gradient(low = "#F3C22D", high = "#3C96DF") +
  theme_classic(base_size = 18) +
  scale_y_continuous(limits = c(-10, 20)) +
  scale_x_continuous(limits = c(NA, 40)) +
  labs(y = "ACC learning score (%)", x = "AQ score", color = "AQ score")

scatter_plot_acc_learning
```

# Figure 5: Model updating as a function of AQ - reaction time
```{r}
# preprocessing
asrt_merged_rt_updating <- asrt_merged_rt %>% 
  select(id, aq_group, aq_sum, bin, ls_binwise_rt) %>% 
  filter(bin == 4 | bin == 5 | bin == 6) %>%
  distinct()
```

## Panel A: line plots with lower and higher AQ groups
```{r}
line_plot_rt_updating <- asrt_merged_rt_updating %>%
  group_by(bin, aq_group) %>% 
  summarise(mean_ls_rt = mean(ls_binwise_rt, na.rm = T), se_ls_rt = std.error(ls_binwise_rt, na.rm = T)) %>%  
  ggplot(aes(bin, mean_ls_rt, color = aq_group)) +
  geom_point() +
  geom_line(linewidth = 1.5) +
  scale_color_manual(values = c("#F3C22D", "#3C96DF")) +
  geom_errorbar(aes(ymin = mean_ls_rt - se_ls_rt, ymax = mean_ls_rt + se_ls_rt), width = 0.1, linewidth = 1) +
  labs(y = "RT learning score (ms)", x = "Bin") +
  scale_x_continuous(breaks = seq(4, 6, 1)) +
  theme_classic(base_size = 18) +
  theme(
    legend.title = element_blank()
  )

line_plot_rt_updating
```

## Panel B: binwise correlations between AQ and learning scores
```{r}
scatter_plot_rt_updating <- asrt_merged_rt_updating %>% 
  ggplot(aes(aq_sum, ls_binwise_rt, color = aq_sum)) +
  geom_jitter() +
  geom_smooth(method = lm, color = "black", lwd = 0.6, alpha = 1) +
  facet_wrap(~ bin) +
  scale_color_gradient(low = "#F3C22D", high = "#3C96DF") +
  theme_classic(base_size = 18) +
  scale_y_continuous(limits = c(-60, 60)) +
  scale_x_continuous(limits = c(NA, 40)) +
  labs(y = "RT learning score (ms)", x = "AQ score", color = "AQ score")

scatter_plot_rt_updating
```

# Figure 6: Model updating as a function of AQ - accuracy
```{r}
asrt_merged_acc_updating <- asrt_merged_acc %>% 
  select(id, aq_sum, aq_group, bin, ls_binwise_acc) %>% 
  filter(bin == 4 | bin == 5 | bin == 6) %>%
  distinct()
```

## Panel A: line plots with lower and higher AQ groups
```{r}
line_plot_acc_updating <- asrt_merged_acc_updating %>%
  group_by(bin, aq_group) %>% 
  summarise(mean_ls_acc = mean(ls_binwise_acc, na.rm = T), se_acc = std.error(ls_binwise_acc, na.rm = T)) %>% 
  ggplot(aes(bin, 100 * mean_ls_acc, color = aq_group)) +
  geom_point() +
  geom_line(linewidth = 1.5) +
  geom_errorbar(aes(ymin = 100 * mean_ls_acc - 100 * se_acc, ymax = 100 * mean_ls_acc + 100 * se_acc), width = 0.1, linewidth = 1) +
  theme_classic(base_size = 18) +
  scale_color_manual(values = c("#F3C22D", "#3C96DF")) +
  theme(
    legend.title = element_blank()
  ) +
  labs(y = "ACC learning score (%)", x = "Bin") +
  scale_x_continuous(breaks = seq(4, 6, 1))

line_plot_acc_updating
```

## Panel B: binwise correlations between AQ and learning scores
```{r}
scatter_plot_acc_updating <- asrt_merged_acc_updating %>% 
  ggplot(aes(aq_sum, 100 * ls_binwise_acc, color = aq_sum)) +
  geom_jitter() +
  geom_smooth(method = lm, color = "black", lwd = 0.6, alpha = 1) +
  facet_wrap(~ bin) +
  scale_color_gradient(low = "#F3C22D", high = "#3C96DF") +
  theme_classic(base_size = 18) +
  scale_y_continuous(limits = c(-10, 20)) +
  scale_x_continuous(limits = c(NA, 40)) +
  labs(y = "ACC learning score (%)", x = "AQ score", color = "AQ score")

scatter_plot_acc_updating
```

# Figure 7: Speed/accuracy trade-off as a function of AQ
```{r}
# preprocessing
mean_sat <- rt_acc_binwise %>%
  group_by(bin, aq_group) %>%
  summarize(mean_sat = mean(bias, na.rm = TRUE), se_sat = std.error(bias, na.rm = TRUE))
```

## Panel A: line plot with lower and higher AQ groups
```{r}
line_plot_sat <-
  ggplot(mean_sat, aes(x = bin, y = mean_sat, group = aq_group, color = aq_group)) +
  geom_errorbar(aes(ymin = mean_sat - se_sat, ymax = mean_sat + se_sat), width = 0.1, linewidth = 1) +
  geom_line(linewidth = 1.5) +
  geom_point() +
  ylim(c(-1.5, 1.5)) +
  theme_classic(base_size = 18) +
  scale_color_manual(values = c("#F3C22D", "#3C96DF")) +
  scale_x_continuous(breaks = seq(1, 6, 1)) +
  theme(legend.position = c(0.75,0.90),
        legend.title = element_blank()
        ) +
  labs(x = "Bin", y = "Speed/accuracy trade-off score")
    
line_plot_sat
```

## Panel B: scatter plot with binwise correlations between AQ and speed/accuracy trade-off
```{r}
scatter_plot_sat <- rt_acc_binwise %>% 
  ggplot(aes(aq_sum, bias, color = aq_sum)) +
  geom_jitter() +
  geom_smooth(method = lm, color = "black", lwd = 0.6, alpha = 1) +
  facet_wrap(~bin) +
  scale_color_gradient(low = "#F3C22D", high = "#3C96DF") +
  theme_classic(base_size = 18) +
  scale_x_continuous(limits = c(NA, 40)) +
  labs(y = "Speed/accuracy trade-off score", x = "AQ score", color = "AQ score")

scatter_plot_sat
```

# Figure S1: Binwise correlations between speed and accuracy
```{r}
# preprocessing
corr_dat_bin <- rt_acc_binwise %>% 
  group_by(bin) %>% 
  summarise(
    bin_corr = cor(bin_median_rt, bin_mean_acc)
  )

rt_acc_binwise <- rt_acc_binwise %>% 
  left_join(corr_dat_bin, by = "bin")

# plotting
scatter_plot_binwise_r <- rt_acc_binwise %>% 
  ggplot(aes(bin_mean_acc*100, bin_median_rt, fill = bin_corr)) +
  geom_point(pch = 21, size = 3.5, color = "black") +
  geom_smooth(method = "lm", color = "black") +
  facet_wrap(. ~ bin) +
  labs(fill = "Speed/accuracy r") +
  xlab("Accuracy (%)") + ylab("RT (ms)") +
  xlim(c(50, 100)) +
  ylim(c(0, 800)) +
  scale_fill_viridis_c(option = "plasma") +
  theme_classic(base_size = 18)
                
scatter_plot_binwise_r
```
